优化图片输出速度
